Build 1.0.0.35
Export:
---Sneak attack Attack, Dice and damage bonus's added to weapon damage export section.
---Fortification Bypass added to weapon damage export section.
UI: Feat selection dialog will now show tooltips during augment selection.
Fix: TWF and TW Blocking feat images sizes now correct.
Fix: Granted feat tooltips will no longer show that their requirements were not met.
Enhancements: Ninja Spy tree updated. Tree version updated to force re-spec on file load.
---It is worth double checking this tree if you use it, as I did not have much time to verify my changes.
Fix: Amounts per AP enhancements will now have correct stack size on file load. (Comparing a number to a string, duh!!!)
Fix: Enhancements will now have correct stack count on file load, not the count of the items tiers. (Wrong value passed in function call)
Fix: Stance stacks counts will now be correct after file load. (Addendum to fix in 1.0.0.32)
Fix: Feat, Enhancement and Destiny effects can now be tagged to be done as Item effects to correctly use stacking rules when required.
---All MovementSpeed effects updated to get correct stacking.
---Note that all items with MovementSpeed bonus's which were of type Enchantment need to be re-equipped to fix.
Fix: A bug in stacking rules was fixed.
Augment Updates:
---Legendary Slavelords Might - Description updated.
---Legendary Slavelords Sorcery - Description updated.
---Legendary Slavelords Endurance - Description updated.
------Resistance Bonus is now Energy Resistance and not a Save bonus.
---All slavelord augments now have icons.
---Meteoric Star Ruby added.
---Ruby Eye augments now have icons.
---All common type augments now have icons.
Item Fixes:
---Beacon of Day - No longer lists Impact V twice.
---Beacon of Day - +2[W] form Impact V now has correct name in breakdown
---Beacon of Night - +2[W] from Impact V now has correct name in breakdown
---Legendary Slave Master's Cloak - "Flase life" text now correctly "False life"
---Epic Arcane Armor - Now has its correct minimum level (26 not 2)
Feat Fixes:
---Strike of the Enduring I - No longer lists +4 to damage all the time as its optional special attack damage.
---Flowing Water Strike I - No longer lists bonus handwraps damage all the time as its optional special attack damage.
---Storm Strike I - No longer lists bonus handwraps damage all the time as its optional special attack damage.
Fix: The City's Beacon Filigree set has returned.
Update 14 Items: (Menace of the Underdark)
---Star of Day (Showtsword)
---Breastplate of the Shining Sun (Medium Armor)
---Implement of the Deep Woods (Quarterstaff)
---Leathers of the Woodsman (Light Armor)
---Oathblade (Longsword)
---Purple Dragon Armor (Heavy Armor)
---Stout Oak Walking Stick (Quarterstaff)
---Vestments of the Sun Soul (Cloth Armor)
---Alarphon's Staff (Quarterstaff)
---War Wizard's Robes (Cloth Armor)
Update 13 Items:
---Epic Cruel Nobility (Scimitar)
---Epic Envenomed Blade (Shortsword)
---Epic Rakshasa Hide (Light Armor)
---Epic Templar's Bulwark (Tower Shield)
---Epic Bow of the Silver Flame (Longbow)
---Epic Flameward (Club)
---Epic Fury of the Flame (Maul)
---Epic Silver Slinger (Repeating Heavy Crossbow)
---Epic Doublecross Bow (Repeating Light Crossbow)
---Epic Luminous Truth (Quarterstaff)
---Epic Templar's Bastion (Medium Armor)
---Epic Templar's Docent (Docent)
---Epic Templar's Justice (Dwarven Axe)
---Epic Templar's Retribution (Bastard Sword)
Update 12 Items:
---None
Update 11 Items:
---Epic Antique Greataxe (Great Axe)
---Epic Kron'zek's Cruelty (Scimitar)
---Epic Phiarlan Spy Dagger (Dagger)
---Epic Noxious Fang (Throwing Dagger)
---Epic Garos' Malice (Shortsword)
---Epic Elyd Edge (Rapier)
---Epic Midnight Greetings (Kukri)

+++ DDOCP/ComboBoxTooltip.cpp
@@ -122,8 +122,8 @@ int CComboBoxTooltip::CompareItem(LPCOMPAREITEMSTRUCT lpCis)
 {
     ASSERT(lpCis->CtlType == ODT_COMBOBOX);
 
-    int iItem1 = (int)(WORD)lpCis->itemID1;
-    int iItem2 = (int)(WORD)lpCis->itemID2;
+    int iItem1 = (int)lpCis->itemID1;
+    int iItem2 = (int)lpCis->itemID2;
     if (iItem1 != CB_ERR && iItem2 != CB_ERR)
     {
         CString strItem1;
